requirements.txt
alembic
docstring_parser==0.7.3
fastapi==0.95.0
FastAPI_SQLAlchemy==0.2.1
passlib==1.7.4
pydantic==1.10.18
PyJWT==2.0.1
psycopg2-binary==2.9.4
pytest==7.3.1
python-dotenv==0.15.0
Requests==2.32.3
SQLAlchemy==2.0.32
starlette==0.26.1
uvicorn==0.21.1
